jira_pptx_app/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jira
lxml==4.9.3
MarkupSafe==2.1.3
Pillow==9.5.0
python-pptx==0.6.22
requests==2.31.0
urllib3==2.0.4
Werkzeug==2.3.7
XlsxWriter==3.1.4
